refactor: 코리 반영(디렉토리 변경)
